Fix overflow error in throughput script.

+++ drv/throughput_plus_heap.d
@@ -1,10 +1,10 @@
 #pragma D option quiet
 
-self int endss;
+self uint64_t endss;
 
 dtrace:::BEGIN
 {
-	allocated = 0;
+	allocated = (uint64_t)0;
 	inter = 0;
 	e = 0;
 }
@@ -18,7 +18,7 @@ syscall::brk:entry
 syscall::brk:entry
 /pid == $target && self->endds < arg0/
 {
-	allocated += ((uint64_t)arg0 - (uint64_t)self->endds);
+	allocated += (uint64_t)((uint64_t)arg0 - (uint64_t)self->endds);
 	/* printf("%d\n", self->allocated); */
 	self->endds = arg0;
 }
